requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
tqdm
torch
torchvision
tensorboard
pyyaml
scipy